Exercise 1: Add AJAX call
In order to do the later exercises it would be easier to not have to reload the page. This change starts that process by introducing an async call to add a new item and manipulating the DOM in the response. Next stage would be to allow an item to be removed